Begin refactor to model the Flask-SQLAlchemy tutorial.
